feat(web): 设置页按功能重组——新增概览落地页与媒体库分组，命名大白话化
分组从 5 组重排为 6 组，标准统一为「回答用户什么问题」：

- 新增「概览」落地页：/settings 直达，未配齐必要件时呈现开局清单
  （三步引导），配齐后是订阅链路体检 + 更新提示卡（复用既有面板与
  待更新快照），问题主动找人；成员不可见，入口按角色兜底到个人信息
- 新增「媒体库」组：刮削与整理迁入；远程转码从「应用」标签升级为
  「播放」分区（按功能命名，给未来播放域设置留位），旧深链
  /settings/app?tab=remote 重定向到 /settings/playback
- 「成员」+「设备」合并为「成员与设备」（同属访问控制），消灭单项组
- 「智能」改组为「通知与集成」：消息推送、Webhook、AI 模型收拢
- 外部访问地址/对外端口从「应用」迁入「网络」分区（网络配置只留一个
  家），app-config-section 拆分为 external-access-section 与
  app-maintenance-section；「应用」瘦身改名「更新与维护」
- 命名清洗：订阅→订阅规则、监听导入→自动入库（含全站用户可见文案，
  后端领域词不动）、网络与代理→网络，描述改为大白话

分区 id 除新增外均未变；订阅页警示横幅改为跳转概览看体检。
