+++ apps/web/components/im-push-section.tsx
@@ -3,7 +3,7 @@
 /**
  * 「消息推送」设置分区。
  *
- * 两类设定按胶囊标签分开（与「应用」「外观」分区同一交互语言），因为它们
+ * 两类设定按胶囊标签分开（与「更新与维护」「外观」分区同一交互语言），因为它们
  * 回答的是两个完全不同的问题，动线不该缠在一起：
  *   - 接入通道：我接了哪些账号？——一张跨平台的统一列表 +「新增通道」菜单，
  *     绑定流程收进弹窗（见 channel-bind-dialog）；
