+++ apps/web/components/library-view.tsx
@@ -1397,13 +1397,13 @@ export function LibraryFormDialog({
             </>
           )}
 
-          {/* 收藏范围（可选）：声明"本库收什么"，订阅与监听导入按它自动选库。
+          {/* 收藏范围（可选）：声明"本库收什么"，订阅与自动入库按它自动选库。
               区域逐国勾选（条件本就是 any_of，预设组降级为一键整组勾选的快捷键）；
               两个维度间是"且" */}
           {tab === "scope" && (
             <>
           <p className="text-sub leading-relaxed text-[var(--text-muted)]">
-            可选：声明「本库收什么」后，订阅与监听导入按作品特征自动选进本库。
+            可选：声明「本库收什么」后，订阅与自动入库按作品特征自动选进本库。
             全部留空 = 不声明，该类型的默认库承接未命中的作品；订阅时永远可以手动改库。
           </p>
           {routingOptions === null ? (
