+++ apps/web/components/player/consent-dialog.tsx
@@ -91,14 +91,14 @@ export function ConsentDialog({
         {canSelfEnable ? (
           <>
             {/* 用户看到这个弹窗的一刻，正是他最需要远程硬件转码的时候——NAS 没有
-                可用硬件编码器，只能拿 CPU 硬扛。这个功能藏在设置三级页里，不在
-                这里说一句就几乎没人会发现。只对能改全局设置的管理员显示：给成员
+                可用硬件编码器，只能拿 CPU 硬扛。这个功能藏在设置的「播放」分区里，
+                不在这里说一句就几乎没人会发现。只对能改全局设置的管理员显示：给成员
                 一个点进去必然 403 的链接毫无意义。 */}
             <p className="mt-4 rounded-xl bg-white/[0.06] px-3 py-2.5 text-[12px] leading-relaxed text-white/60">
               局域网里有 Apple Silicon Mac 的话，可以让它替 NAS 做硬件转码，省下这里的
               CPU 开销。
               <Link
-                href={{ pathname: "/settings/app", query: { tab: "remote" } }}
+                href={{ pathname: "/settings/playback" }}
                 className="ml-1 text-[var(--player-accent)] underline decoration-dotted underline-offset-2"
               >
                 去设置远程转码
