+++ apps/web/components/remote-transcode-section.tsx
@@ -44,7 +44,8 @@ export interface RemoteTranscodeSectionProps {
 }
 
 /**
- * 「应用 → 远程转码」设置。
+ * 「播放」分区的远程转码设置（原「应用 → 远程转码」标签，设置页按功能重组后
+ * 迁入「媒体库」组）。
  *
  * 这一页要人做的决定只剩一个：开还是不开。
  *
