+++ apps/web/components/rule-sets-panel.tsx
@@ -24,7 +24,7 @@ import {
 } from "@/lib/upgrade-ladder";
 
 /**
- * 规则组管理（设置 → 订阅 → 规则组）：Web 端唯一的规则组配置入口。
+ * 规则组管理（设置 → 订阅规则 → 规则组）：Web 端唯一的规则组配置入口。
  *
  * 设计原则：
  * - **表单化，不让用户手写 JSON**——spec 的每个维度铺成对应控件；
