+++ apps/web/components/upgrade-run-dialog.tsx
@@ -169,7 +169,7 @@ export function UpgradeRunDialog({
                       + 新建规则组
                     </button>
                   ) : (
-                    " 请联系管理员在「设置 → 订阅 → 规则组」中配置「洗到哪一档」。"
+                    " 请联系管理员在「设置 → 订阅规则 → 规则组」中配置「洗到哪一档」。"
                   )}
                 </div>
               ) : (
